markdownlint: add -g option to check_format and fix mdl errors
-g option using only files known to git, that allow us
to avoid errors from submodules

Also add check if mdl is installed, and gracefull info if not.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/9086
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ scripts/check_format.sh
@@ -587,16 +587,21 @@ function check_json_rpc() {
 function check_markdown_format() {
 	local rc=0
 
-	echo -n "Checking markdown files format..."
-	mdl -s $rootdir/mdl_rules.rb . > mdl.log || true
-	if [ -s mdl.log ]; then
-		echo " Errors in .md files detected:"
-		cat mdl.log
-		rc=1
+	if hash mdl 2> /dev/null; then
+		echo -n "Checking markdown files format..."
+		mdl -g -s $rootdir/mdl_rules.rb . > mdl.log || true
+		if [ -s mdl.log ]; then
+			echo " Errors in .md files detected:"
+			cat mdl.log
+			rc=1
+		else
+			echo " OK"
+		fi
+		rm -f mdl.log
 	else
-		echo " OK"
+		echo "You do not have markdownlint installed so .md files not being checked!"
 	fi
-	rm -f mdl.log
+
 	return $rc
 }
